requirements.txt
pika==1.3.2
python-dotenv>=1.0.0
simple-salesforce>=1.12.0
requests>=2.31.0
